remove deprecated providers
