+++ src/MigrationServiceProvider.php
@@ -2,6 +2,8 @@
 
 namespace Vinelab\NeoEloquent;
 
+use Illuminate\Support\ServiceProvider;
+use Illuminate\Database\Migrations\Migrator;
 use Vinelab\NeoEloquent\Migrations\MigrationModel;
 use Vinelab\NeoEloquent\Migrations\MigrationCreator;
 use Vinelab\NeoEloquent\Console\Migrations\MigrateCommand;
@@ -10,11 +12,9 @@ use Vinelab\NeoEloquent\Console\Migrations\MigrateResetCommand;
 use Vinelab\NeoEloquent\Migrations\DatabaseMigrationRepository;
 use Vinelab\NeoEloquent\Console\Migrations\MigrateRefreshCommand;
 use Vinelab\NeoEloquent\Console\Migrations\MigrateRollbackCommand;
-use Illuminate\Database\Migrations\Migrator;
-use Illuminate\Support\ServiceProvider;
 
-class MigrationServiceProvider extends ServiceProvider
-{
+class  MigrationServiceProvider extends ServiceProvider {
+
     /**
      * {@inheritDoc}
      */
@@ -44,11 +44,14 @@ class MigrationServiceProvider extends ServiceProvider
 
     /**
      * Register the migration repository service.
+     *
+     * @return void
      */
     protected function registerRepository()
     {
-        $this->app->bindShared('neoeloquent.migration.repository', function ($app) {
-            $model = new MigrationModel();
+        $this->app->singleton('neoeloquent.migration.repository', function($app)
+        {
+            $model = new MigrationModel;
 
             $label = $app['config']['database.migrations_node'];
 
@@ -66,21 +69,26 @@ class MigrationServiceProvider extends ServiceProvider
 
     /**
      * Register the migrator service.
+     *
+     * @return void
      */
     protected function registerMigrator()
     {
         // The migrator is responsible for actually running and rollback the migration
         // files in the application. We'll pass in our database connection resolver
         // so the migrator can resolve any of these connections when it needs to.
-        $this->app->bindShared('neoeloquent.migrator', function ($app) {
+        $this->app->singleton('neoeloquent.migrator', function($app) {
             $repository = $app['neoeloquent.migration.repository'];
 
             return new Migrator($repository, $app['db'], $app['files']);
         });
     }
 
+
     /**
      * Register all of the migration commands.
+     *
+     * @return void
      */
     protected function registerCommands()
     {
@@ -89,13 +97,14 @@ class MigrationServiceProvider extends ServiceProvider
             'MigrateRollback',
             'MigrateReset',
             'MigrateRefresh',
-            'MigrateMake',
+            'MigrateMake'
         );
 
         // We'll simply spin through the list of commands that are migration related
         // and register each one of them with an application container. They will
         // be resolved in the Artisan start file and registered on the console.
-        foreach ($commands as $command) {
+        foreach ($commands as $command)
+        {
             $this->{'register'.$command.'Command'}();
         }
 
@@ -113,10 +122,12 @@ class MigrationServiceProvider extends ServiceProvider
 
     /**
      * Register the "migrate" migration command.
+     *
+     * @return void
      */
     protected function registerMigrateCommand()
     {
-        $this->app->bindShared('command.neoeloquent.migrate', function ($app) {
+        $this->app->singleton('command.neoeloquent.migrate', function($app) {
             $packagePath = $app['path.base'].'/vendor';
 
             return new MigrateCommand($app['neoeloquent.migrator'], $packagePath);
@@ -125,44 +136,55 @@ class MigrationServiceProvider extends ServiceProvider
 
     /**
      * Register the "rollback" migration command.
+     *
+     * @return void
      */
     protected function registerMigrateRollbackCommand()
     {
-        $this->app->bindShared('command.neoeloquent.migrate.rollback', function ($app) {
+        $this->app->singleton('command.neoeloquent.migrate.rollback', function($app)
+        {
             return new MigrateRollbackCommand($app['neoeloquent.migrator']);
         });
     }
 
     /**
      * Register the "reset" migration command.
+     *
+     * @return void
      */
     protected function registerMigrateResetCommand()
     {
-        $this->app->bindShared('command.neoeloquent.migrate.reset', function ($app) {
+        $this->app->singleton('command.neoeloquent.migrate.reset', function($app)
+        {
             return new MigrateResetCommand($app['neoeloquent.migrator']);
         });
     }
 
     /**
      * Register the "refresh" migration command.
+     *
+     * @return void
      */
     protected function registerMigrateRefreshCommand()
     {
-        $this->app->bindShared('command.neoeloquent.migrate.refresh', function ($app) {
+        $this->app->singleton('command.neoeloquent.migrate.refresh', function($app)
+        {
             return new MigrateRefreshCommand();
         });
     }
 
     /**
      * Register the "install" migration command.
+     *
+     * @return void
      */
     protected function registerMigrateMakeCommand()
     {
-        $this->app->bindShared('migration.neoeloquent.creator', function ($app) {
+        $this->app->singleton('migration.neoeloquent.creator', function($app) {
             return new MigrationCreator($app['files']);
         });
 
-        $this->app->bindShared('command.neoeloquent.migrate.make', function ($app) {
+        $this->app->singleton('command.neoeloquent.migrate.make', function($app) {
             // Once we have the migration creator registered, we will create the command
             // and inject the creator. The creator is responsible for the actual file
             // creation of the migrations, and may be extended by these developers.
@@ -170,7 +192,7 @@ class MigrationServiceProvider extends ServiceProvider
 
             $packagePath = $app['path.base'].'/vendor';
 
-            $composer = $app->make('Illuminate\Foundation\Composer');
+            $composer = $app->make('Illuminate\Support\Composer');
 
             return new MigrateMakeCommand($creator, $composer, $packagePath);
         });
@@ -192,4 +214,5 @@ class MigrationServiceProvider extends ServiceProvider
             'command.neoeloquent.migrate.make',
         );
     }
+
 }
