+++ src/NeoEloquentServiceProvider.php
@@ -4,6 +4,10 @@ namespace Vinelab\NeoEloquent;
 
 use Vinelab\NeoEloquent\Eloquent\Model;
 use Vinelab\NeoEloquent\Schema\Grammars\CypherGrammar;
+use Vinelab\NeoEloquent\Connection as NeoEloquentConnection;
+use Vinelab\NeoEloquent\Frameworks\Laravel\Connection;
+
+use Illuminate\Events\Dispatcher;
 use Illuminate\Support\ServiceProvider;
 
 class NeoEloquentServiceProvider extends ServiceProvider
@@ -31,7 +35,7 @@ class NeoEloquentServiceProvider extends ServiceProvider
     {
         Model::setConnectionResolver($this->app['db']);
 
-        Model::setEventDispatcher($this->app['events']);
+        Model::setEventDispatcher($this->app->make(Dispatcher::class));
     }
 
     /**
@@ -40,7 +44,18 @@ class NeoEloquentServiceProvider extends ServiceProvider
     public function register()
     {
         $this->app['db']->extend('neo4j', function ($config) {
+            $this->config = $config;
             $conn = new Connection($config);
+            $conn->setSchemaGrammar(new CypherGrammar());
+
+            return $conn;
+        });
+
+        $this->app->bind('neoeloquent.connection', function() {
+            // $config is set by the previous binding,
+            // so that we get the correct configuration
+            // set by the user.
+            $conn = new NeoEloquentConnection($this->config);
             $conn->setSchemaGrammar(new CypherGrammar());
 
             return $conn;
@@ -53,6 +68,16 @@ class NeoEloquentServiceProvider extends ServiceProvider
         });
 
         $this->registerComponents();
+    }
+
+    protected function registerNeoEloquentConnection($app, $config)
+    {
+        $app->bind('neoeloquent.connection', function() use($config) {
+            $conn = new NeoEloquentConnection($config);
+            $conn->setSchemaGrammar(new CypherGrammar());
+
+            return $conn;
+        });
     }
 
     /**
@@ -72,7 +97,7 @@ class NeoEloquentServiceProvider extends ServiceProvider
      */
     protected function registerMigration()
     {
-        $this->app->register('Vinelab\NeoEloquent\MigrationServiceProvider');
+        $this->app->register(MigrationServiceProvider::class);
     }
 
     /**
